+++ examples/web/src/pages/StartPage.vue
@@ -68,7 +68,7 @@ import PageLink from '../components/PageLink.vue'
     <div class="page-links">
       <PageLink href="http://localhost:5062/scalar">
         <template #title>Nuxt</template>
-        <template #description>@scalar/nuxt</template>
+        <template #description>@readyapi/nuxt</template>
       </PageLink>
       <PageLink href="http://localhost:5063/scalar">
         <template #title>Docusaurus</template>