+++ packages/api-reference/src/components/Content/Authentication/Authentication.vue
@@ -5,7 +5,7 @@ import {
   useAuthenticationStore,
 } from '@readyapi/api-client'
 import { type SSRState } from '@readyapi/oas-utils'
-import type { OpenAPIV3_1 } from '@scalar/openapi-parser'
+import type { OpenAPIV3_1 } from '@readyapi/openapi-parser'
 import { computed, onServerPrefetch, useSSRContext, watch } from 'vue'
 
 import { hasSecuritySchemes, sleep } from '../../../helpers'