+++ packages/api-reference/src/components/Content/Introduction/Introduction.vue
@@ -3,7 +3,7 @@ import {
   type OpenAPIV2,
   type OpenAPIV3,
   type OpenAPIV3_1,
-} from '@scalar/openapi-parser'
+} from '@readyapi/openapi-parser'
 import { computed } from 'vue'
 
 import type { Spec } from '../../../types'