+++ packages/api-reference/src/components/Content/Schema/SchemaHeading.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import { type OpenAPIV2 } from '@scalar/openapi-parser'
+import { type OpenAPIV2 } from '@readyapi/openapi-parser'
 
 defineProps<{
   value: OpenAPIV2.SchemaObject