+++ packages/cli/src/utils/loadOpenApiFile.ts
@@ -1,4 +1,4 @@
-import { openapi } from '@scalar/openapi-parser'
+import { openapi } from '@readyapi/openapi-parser'
 import kleur from 'kleur'
 
 import { getFileOrUrl } from './getFileOrUrl'