+++ packages/api-client-react/CHANGELOG.md
@@ -224,7 +224,7 @@
 
 ### Minor Changes
 
-- 7fb8273: Migrate to @scalar/openapi-parser
+- 7fb8273: Migrate to @readyapi/openapi-parser
 
 ### Patch Changes