+++ packages/echo-server/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Minor Changes
 
-- 7fb8273: Migrate to @scalar/openapi-parser
+- 7fb8273: Migrate to @readyapi/openapi-parser
 
 ## 0.5.14