+++ packages/fastify-api-reference/CHANGELOG.md
@@ -124,7 +124,7 @@
 
 ### Minor Changes
 
-- 7fb8273: Migrate to @scalar/openapi-parser
+- 7fb8273: Migrate to @readyapi/openapi-parser
 
 ## 1.18.7