+++ packages/mock-server/CHANGELOG.md
@@ -374,7 +374,7 @@
 
 ### Minor Changes
 
-- a866e91: refactor: switch to @scalar/openapi-parser
+- a866e91: refactor: switch to @readyapi/openapi-parser
 
 ### Patch Changes