+++ packages/nestjs-api-reference/CHANGELOG.md
@@ -423,7 +423,7 @@
 
 ### Minor Changes
 
-- 7fb8273: Migrate to @scalar/openapi-parser
+- 7fb8273: Migrate to @readyapi/openapi-parser
 
 ### Patch Changes