+++ packages/nextjs-api-reference/CHANGELOG.md
@@ -425,7 +425,7 @@
 
 ### Minor Changes
 
-- 7fb8273: Migrate to @scalar/openapi-parser
+- 7fb8273: Migrate to @readyapi/openapi-parser
 
 ### Patch Changes