+++ packages/oas-utils/CHANGELOG.md
@@ -40,7 +40,7 @@
 
 ### Minor Changes
 
-- 7fb8273: Migrate to @scalar/openapi-parser
+- 7fb8273: Migrate to @readyapi/openapi-parser
 
 ## 0.0.4