+++ packages/use-codemirror/CHANGELOG.md
@@ -26,7 +26,7 @@
 
 ### Minor Changes
 
-- 7fb8273: Migrate to @scalar/openapi-parser
+- 7fb8273: Migrate to @readyapi/openapi-parser
 
 ## 0.8.10